backend_python/gloss_converter.py
"""
Gloss Converter Module
Converts text to gloss format for sign language translation.
"""


# Global variables for gloss model (lazy loading)
gloss_model = None
gloss_tokenizer = None


def load_gloss_model():
    """
    Load the gloss translation model.
    This is a placeholder until the gloss_translator package is properly installed.
    
    Returns:
        tuple: (model, tokenizer) or (None, None) if unavailable
    """
    global gloss_model, gloss_tokenizer
    
    if gloss_model is not None:
        return gloss_model, gloss_tokenizer
    
    try:
        # Try to import the gloss_translator package
        from gloss_translator import GlossTranslator
        
        print("Loading gloss translation model...")
        translator = GlossTranslator()
        gloss_model = translator
        gloss_tokenizer = translator  # Placeholder
        print("Gloss model loaded successfully")
        
        return gloss_model, gloss_tokenizer
        
    except ImportError:
        print("gloss_translator package not available, using placeholder")
        return None, None
    except Exception as e:
        print(f"Error loading gloss model: {str(e)}")
        return None, None


def translate_to_gloss_format(text):
    """
    Translate text to gloss format (sign language notation).
    
    Args:
        text (str): Text to translate
    
    Returns:
        str: Text in gloss format
    """
    try:
        model, tokenizer = load_gloss_model()
        
        if model is None:
            # Placeholder translation until gloss_translator is available
            # Convert to uppercase and basic simplification
            gloss_text = text.upper()
            gloss_text = gloss_text.replace("'", "")
            gloss_text = ' '.join(gloss_text.split())
            
            return f"[GLOSS TRANSLATION - PLACEHOLDER MODE]\n\n{gloss_text}\n\n[Note: Install gloss_translator for accurate translations]"
        
        # Use actual gloss translator
        gloss_translation = model.translate(text)
        return gloss_translation
        
    except Exception as e:
        print(f"Error translating to gloss: {str(e)}")
        import traceback
        traceback.print_exc()
        
        # Return placeholder on error
        return f"[GLOSS TRANSLATION ERROR]\n\nOriginal text:\n{text}\n\n[Error: {str(e)}]"


def convert_topics_to_gloss(topics):
    """
    Convert a list of topic sections to gloss format.
    
    Args:
        topics (list): List of topic dictionaries
    
    Returns:
        list: Topics with added 'gloss' field
    """
    try:
        for topic in topics:
            if 'text' in topic:
                topic['gloss'] = translate_to_gloss_format(topic['text'])
        
        return topics
        
    except Exception as e:
        print(f"Error converting topics to gloss: {str(e)}")
        return topics
