backend_python/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
youtube-transcript-api==0.6.1
yt-dlp==2023.12.30
python-dotenv==1.0.0
gunicorn==21.2.0
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.26.2
torch>=1.9.0
transformers>=4.18.0
jiwer>=2.3.0
